backend/requirements.txt
langchain==1.0.0a9
langchain-openai==0.3.33
python-dotenv==1.1.1
fastapi==0.117.1
uvicorn==0.37.0
loguru==0.7.3
pydantic_settings==2.11.0

Pillow==11.3.0
PyMuPDF==1.26.4
python-magic-bin==0.4.14
langchain_community==0.3.30

langchain-unstructured==0.1.6
langchain-text-splitters==0.3.11
unstructured[pdf]==0.18.15
matplotlib==3.8.0

# OCR and PDF processing dependencies
pytesseract>=0.3.10
pdf2image>=1.17.0
# Note: Also requires system Tesseract OCR installation
# macOS: brew install tesseract
# Ubuntu: sudo apt-get install tesseract-ocr
# Windows: https://github.com/UB-Mannheim/tesseract/wiki

# Audio processing
pydub>=0.25.1
moviepy==1.0.3
# Note: Also requires system ffmpeg installation
# macOS: brew install ffmpeg
# Ubuntu: sudo apt-get install ffmpeg
# Windows: https://ffmpeg.org/download.html